+ [3.0-RC-3] Update TabooLib to 6.0.4-3.   * Fixed #88

+++ src/main/kotlin/me/arasple/mc/trmenu/module/internal/database/DatabaseLocal.kt
@@ -24,7 +24,7 @@ class DatabaseLocal : Database() {
         val origin = File(getDataFolder(), "data.db")
 
         // 迁移到新目录
-        if (origin.exists()) {
+        if (origin.exists() && !target.exists()) {
             origin.copyTo(target)
             origin.delete()
         }
